Updated failing execution tests related to keccak256, sha256, ripemd160. Used the tests to verify the relevant issue (#args, padding).

+++ test/libsolidity/semanticTests/builtinFunctions/keccak256_multiple_arguments_with_numeric_literals.sol
@@ -1,9 +1,9 @@
 contract c {
-    function foo(uint a, uint16 b) public returns (bytes32 d) {
+    function foo(uint256 a, uint16 b) public returns (bytes32 d) {
         d = keccak256(abi.encodePacked(a, b, uint8(145)));
     }
 }
 // ====
 // compileViaYul: also
 // ----
-// foo(uint,uint16): 0xa, 0xc -> 0x0088acd45f75907e7c560318bc1a5249850a0999c4896717b1167d05d116e6dbad
+// foo(uint256,uint16): 0x0a, 0x0c -> 0x0088acd45f75907e7c560318bc1a5249850a0999c4896717b1167d05d116e6dbad
